rotation wip

+++ library/ui/desktop_ui/src/CheckBox.cpp
@@ -15,7 +15,7 @@ void RadioGroup::set_active(const CheckBox* activeCb) const
 {
    const auto it = std::ranges::find(m_checkBoxes, activeCb);
    if (it != m_checkBoxes.end()) {
-      event_OnSelection.publish(it - m_checkBoxes.begin());
+      event_OnSelection.publish(static_cast<u32>(it - m_checkBoxes.begin()));
    }
 
    for (CheckBox* checkBox : m_checkBoxes) {
